# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Aimeos Store API",
    "version": "1.0.0",
    "description": "REST API endpoints for Aimeos storefront user registration, authentication, and profile management. Authorize using a Bearer token generated on login/register."
  },
  "servers": [
    {
      "url": "/",
      "description": "Current Host Server"
    }
  ],
  "paths": {
    "/api/register/customer": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Register a new Customer",
        "description": "Register a new Customer account on the default site. Returns a Sanctum authentication token.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "The user's display name",
                    "example": "Jane Doe"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "A unique email address",
                    "example": "customer@example.com"
                  },
                  "password": {
                    "type": "string",
                    "format": "password",
                    "description": "User password (minimum 8 characters)",
                    "example": "password123"
                  },
                  "password_confirmation": {
                    "type": "string",
                    "format": "password",
                    "description": "Must match the password field",
                    "example": "password123"
                  }
                },
                "required": [
                  "name",
                  "email",
                  "password",
                  "password_confirmation"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Registration successful. Returns user details and API token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed (e.g. email already exists, mismatched password).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/register/seller": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Register a new Seller",
        "description": "Register a new Seller account, creating a new tenant/site with the specified code. Returns a Sanctum authentication token.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "code": {
                    "type": "string",
                    "description": "The unique site/tenant code to create",
                    "example": "sellershop"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "A unique email address",
                    "example": "seller@example.com"
                  },
                  "password": {
                    "type": "string",
                    "format": "password",
                    "description": "User password (minimum 8 characters)",
                    "example": "password123"
                  },
                  "password_confirmation": {
                    "type": "string",
                    "format": "password",
                    "description": "Must match the password field",
                    "example": "password123"
                  }
                },
                "required": [
                  "code",
                  "email",
                  "password",
                  "password_confirmation"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Registration successful. Returns user details and API token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed (e.g. email already exists, duplicate site code).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/register/admin": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Register a new Admin",
        "description": "Register a new Administrator account on the default site. Returns a Sanctum authentication token.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "The user's display name",
                    "example": "Admin User"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "A unique email address",
                    "example": "admin@example.com"
                  },
                  "password": {
                    "type": "string",
                    "format": "password",
                    "description": "User password (minimum 8 characters)",
                    "example": "password123"
                  },
                  "password_confirmation": {
                    "type": "string",
                    "format": "password",
                    "description": "Must match the password field",
                    "example": "password123"
                  }
                },
                "required": [
                  "name",
                  "email",
                  "password",
                  "password_confirmation"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Registration successful. Returns user details and API token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed (e.g. email already exists, mismatched password).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/login": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Authenticate / Log in",
        "description": "Authenticate user credentials and issue a Sanctum token.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "jane.doe@example.com"
                  },
                  "password": {
                    "type": "string",
                    "format": "password",
                    "example": "password123"
                  }
                },
                "required": [
                  "email",
                  "password"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authentication successful.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed or invalid credentials.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/logout": {
      "post": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Authentication"
        ],
        "summary": "Log out / Revoke Token",
        "description": "Revoke the access token currently used to authenticate the request.",
        "responses": {
          "200": {
            "description": "Successful logout.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Logged out successfully"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized / Missing or invalid token."
          }
        }
      }
    },
    "/api/profile": {
      "get": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Profile"
        ],
        "summary": "Get Profile details",
        "description": "Retrieve the current authenticated user's model.",
        "responses": {
          "200": {
            "description": "User profile retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized / Missing or invalid token."
          }
        }
      },
      "put": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Profile"
        ],
        "summary": "Update Profile",
        "description": "Update name or email address of the authenticated user.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Jane Doe Updated"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "jane.updated@example.com"
                  }
                },
                "required": [
                  "name",
                  "email"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profile updated successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Profile updated successfully"
                    },
                    "user": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation errors.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized."
          }
        }
      },
      "delete": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Profile"
        ],
        "summary": "Delete Account",
        "description": "Delete the authenticated user's account permanently. Requires confirming current password.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "password": {
                    "type": "string",
                    "format": "password",
                    "description": "The user's current password for verification",
                    "example": "password123"
                  }
                },
                "required": [
                  "password"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Account deleted successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Account deleted successfully"
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation error or incorrect password.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized."
          }
        }
      }
    },
    "/api/password": {
      "put": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Profile"
        ],
        "summary": "Update Password",
        "description": "Change the authenticated user's password.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "current_password": {
                    "type": "string",
                    "format": "password",
                    "example": "password123"
                  },
                  "password": {
                    "type": "string",
                    "format": "password",
                    "description": "The new password",
                    "example": "newsecurepass99"
                  },
                  "password_confirmation": {
                    "type": "string",
                    "format": "password",
                    "description": "Must match the new password field",
                    "example": "newsecurepass99"
                  }
                },
                "required": [
                  "current_password",
                  "password",
                  "password_confirmation"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password updated successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Password updated successfully"
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation error (e.g. incorrect current password or password validation failure).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized."
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Jane Doe"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "jane.doe@example.com"
          },
          "email_verified_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "example": "2026-05-21T12:00:00.000000Z"
          },
          "siteid": {
            "type": "string",
            "example": "1."
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "example": "2026-05-21T11:00:00.000000Z"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "example": "2026-05-21T12:00:00.000000Z"
          }
        }
      },
      "AuthResponse": {
        "type": "object",
        "properties": {
          "access_token": {
            "type": "string",
            "example": "1|abc123xyzTokenValue..."
          },
          "token_type": {
            "type": "string",
            "example": "Bearer"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          }
        }
      },
      "ValidationError": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "The email has already been taken."
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "example": {
              "email": [
                "The email has already been taken."
              ]
            }
          }
        }
      }
    }
  }
}